+++ tracker/tracker/src/main/modules/analytics/events.ts
@@ -1,16 +1,11 @@
-import Batcher from './batcher.js';
+import Batcher from './batcher.js'
 import ConstantProperties from './constantProperties.js'
 import { isObject } from './utils.js'
-import { createEvent, categories } from './types.js';
+import { createEvent, categories } from './types.js'
 
-const maxProperties = 100;
-const maxPropLength = 100;
-
-const reservedProps = [
-  'properties',
-  'token',
-  'timestamp',
-];
+const maxProperties = 100
+const maxPropLength = 100
+const reservedProps = ['properties', 'token', 'timestamp']
 
 export default class Events {
   ownProperties: Record<string, any> = {}
@@ -20,7 +15,7 @@ export default class Events {
     private readonly getTimestamp: () => number,
     private readonly batcher: Batcher,
   ) {
-    this.ownProperties = this.constantProperties.getSuperProperties();
+    this.ownProperties = this.constantProperties.getSuperProperties()
   }
 
   /**
@@ -48,16 +43,11 @@ export default class Events {
       name: eventName,
       properties: { ...this.ownProperties, ...eventProps },
     }
-    const event = createEvent(
-      categories.events,
-      undefined,
-      this.getTimestamp(),
-      eventPayload,
-    )
+    const event = createEvent(categories.events, undefined, this.getTimestamp(), eventPayload)
     if (options?.send_immediately) {
-      void this.batcher.sendImmediately(event);
+      void this.batcher.sendImmediately(event)
     } else {
-      this.batcher.addEvent(event);
+      this.batcher.addEvent(event)
     }
   }
 
@@ -65,24 +55,24 @@ export default class Events {
    * creates super property for all events
    * */
   setProperty = (nameOrProperties: Record<string, any> | string, value?: any) => {
-    let changed = false;
+    let changed = false
     if (isObject(nameOrProperties)) {
       Object.entries(nameOrProperties).forEach(([key, val]) => {
         if (!this.constantProperties.defaultPropertyKeys.includes(key)) {
           this.ownProperties[key] = val
-          changed = true;
+          changed = true
         }
       })
     }
     if (typeof nameOrProperties === 'string' && value !== undefined) {
       if (!this.constantProperties.defaultPropertyKeys.includes(nameOrProperties)) {
         this.ownProperties[nameOrProperties] = value
-        changed = true;
+        changed = true
       }
     }
 
     if (changed) {
-      this.constantProperties.saveSuperProperties(this.ownProperties);
+      this.constantProperties.saveSuperProperties(this.ownProperties)
     }
   }
 
@@ -90,24 +80,24 @@ export default class Events {
    * set super property only if it doesn't exist yet
    * */
   setPropertiesOnce = (nameOrProperties: Record<string, any> | string, value?: any) => {
-    let changed = false;
+    let changed = false
     if (isObject(nameOrProperties)) {
       Object.entries(nameOrProperties).forEach(([key, val]) => {
         if (!this.ownProperties[key] && !reservedProps.includes(key)) {
           this.ownProperties[key] = val
-          changed = true;
+          changed = true
         }
       })
     }
     if (typeof nameOrProperties === 'string' && value !== undefined) {
       if (!this.ownProperties[nameOrProperties] && !reservedProps.includes(nameOrProperties)) {
         this.ownProperties[nameOrProperties] = value
-        changed = true;
+        changed = true
       }
     }
 
     if (changed) {
-      this.constantProperties.saveSuperProperties(this.ownProperties);
+      this.constantProperties.saveSuperProperties(this.ownProperties)
     }
   }
 
@@ -115,27 +105,27 @@ export default class Events {
    * removes properties from list of super properties
    * */
   unsetProperties = (properties: string | string[]) => {
-    let changed = false;
+    let changed = false
     if (Array.isArray(properties)) {
       properties.forEach((key) => {
         if (this.ownProperties[key] && !reservedProps.includes(key)) {
           delete this.ownProperties[key]
-          changed = true;
+          changed = true
         }
       })
     } else if (this.ownProperties[properties] && !reservedProps.includes(properties)) {
       delete this.ownProperties[properties]
-      changed = true;
+      changed = true
     }
     if (changed) {
-      this.constantProperties.saveSuperProperties(this.ownProperties);
+      this.constantProperties.saveSuperProperties(this.ownProperties)
     }
   }
 
   /** clears all super properties */
   reset = () => {
     this.ownProperties = {}
-    this.constantProperties.clearSuperProperties();
+    this.constantProperties.clearSuperProperties()
   }
 
   /** mixpanel compatibility */